requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-cors-headers>=4.7.0
whitenoise>=6.6.0
pillow
# API Documentation
drf-yasg>=1.21.7
Pygments>=2.13.0
coreapi>=2.3.3
uritemplate>=4.1.1